pyproject.toml
[project]
name = "protein-toolbox"
version = "0.1.0"
description = "A suite of protein design tools with agentic workflow support."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "biopython",
    "openmm",
    "crewai",
    "boltz",
    "pandas",
    "requests",
    "streamlit",
    "stmol",
    "py3Dmol",
    "streamlit-molstar",
    "vina",
    "protpy",
    "meeko",
    "freesasa"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/proteintoolbox"]
